Removed a lot of code duplication regarding razor pages.
I basically just tried to use inheritance and partial views to extract almost all common code. Tbh, it's not perfect and it kinda does suck in some ways, but it's the best I can do.
