+++ src/Chirp.Razor/Pages/Shared/_CheepPostBox.cshtml
@@ -1,0 +1,20 @@
+﻿@model Chirp.Razor.Pages.Shared.CheepTimeline
+
+<div class="cheepbox" style="min-height: 60px">
+    <h3>What's on your mind @(User.Identity?.Name)?</h3>
+    <form method="post">
+        <input id="cheepText" style="float: left" type="text" asp-for="Text">
+        <p id="charCount" style="float: left">160 characters remaining</p>
+        <input id="btnShare" type="submit" value="Share" disabled>
+    </form>
+</div>
+<script>
+	document.getElementById('cheepText').addEventListener('input', function () {
+	    const maxLength = 160;
+		const remaining = maxLength - this.value.length;
+		document.getElementById('charCount').innerText = remaining + ' characters remaining';
+
+		// https://stackoverflow.com/a/13831737
+		document.getElementById("btnShare").disabled = remaining === 160;
+	});	
+</script>
