+++ src/Chirp.Razor/Pages/UserTimeline.cshtml
@@ -8,91 +8,5 @@
 
 <div>
     <h2> @routeName's Timeline </h2>
-    @if (User.Identity is { IsAuthenticated: true })
-	{
-		<div class="cheepbox" style="min-height: 60px">
-			<h3>What's on your mind @(User.Identity.Name)?</h3>
-			<form method="post">
-				<input id="cheepText" style="float: left" type="text" asp-for="Text">
-				<p id="charCount" style="float: left">160 characters remaining</p>
-				<input id="btnShare" type="submit" value="Share" disabled>
-			</form>
-		</div>
-		<script>
-			document.getElementById('cheepText').addEventListener('input', function () {
-				const maxLength = 160;
-				const remaining = maxLength - this.value.length;
-				document.getElementById('charCount').innerText = remaining + ' characters remaining';
-
-				// https://stackoverflow.com/a/13831737
-				document.getElementById("btnShare").disabled = remaining === 160;
-			});
-		</script>
-	}
-	@if (Model.Cheeps.Any())
-    {
-        <ul id="messagelist" class="cheeps">
-            @foreach (var cheep in Model.Cheeps)
-            {
-                <li>
-					<p style="word-wrap:break-word;">
-                        <strong>
-                            <a href="/@cheep.Author">@cheep.Author</a>
-                        </strong>
-                        @cheep.Message
-                        <small>&mdash; @cheep.Timestamp</small>
-                    </p>
-                </li>
-            }
-        </ul>
-
-        <nav aria-label="Page navigation">
-            <ul style="list-style-type: none; padding: 0; display: flex; justify-content: center;">
-                @if (Model.CurrentPage > 1)
-                {
-                    <li style="margin: 0 5px;">
-                        <a href="/@Model.RouteData.Values["author"]?page=@(Model.CurrentPage - 1)" aria-label="Previous" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">&laquo;</a>
-                    </li>
-                }
-
-                @if (Model.CurrentPage > 3)
-                {
-                    <li style="margin: 0 5px;">
-                        <a href="/@Model.RouteData.Values["author"]?page=1" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">1</a>
-                    </li>
-                    <li style="margin: 0 5px;">
-                        <span style="padding: 0.5em 1em; border: 1px solid #ddd;">...</span>
-                    </li>
-                }
-
-                @for (int i = Math.Max(1, Model.CurrentPage - 2); i <= Math.Min(Model.TotalPages, Model.CurrentPage + 2); i++)
-                {
-                    <li style="margin: 0 5px;" class="@(Model.CurrentPage == i ? "active" : "")">
-                        <a href="/@Model.RouteData.Values["author"]?page=@i" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">@i</a>
-                    </li>
-                }
-
-                @if (Model.CurrentPage < Model.TotalPages - 2)
-                {
-                    <li style="margin: 0 5px;">
-                        <span style="padding: 0.5em 1em; border: 1px solid #ddd;">...</span>
-                    </li>
-                    <li style="margin: 0 5px;">
-                        <a href="/@Model.RouteData.Values["author"]?page=@Model.TotalPages" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">@Model.TotalPages</a>
-                    </li>
-                }
-
-                @if (Model.CurrentPage < Model.TotalPages)
-                {
-                    <li style="margin: 0 5px;">
-                        <a href="/@Model.RouteData.Values["author"]?page=@(Model.CurrentPage + 1)" aria-label="Next" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">&raquo;</a>
-                    </li>
-                }
-            </ul>
-        </nav>
-    }
-    else
-    {
-        <em>There are no cheeps so far.</em>
-    }
+    <partial name="Shared/_CheepTimeline"/>
 </div>
